Use 64-bit, NME_ANGLE for acadnme.

+++ project/src/windows/System.cpp
@@ -28,7 +28,7 @@ namespace nme {
    bool LaunchBrowser(const char *inUtf8URL)
    {
       int result;
-      result=(int)ShellExecute(NULL, "open", inUtf8URL, NULL, NULL, SW_SHOWDEFAULT);
+      result=(int)(size_t)ShellExecute(NULL, "open", inUtf8URL, NULL, NULL, SW_SHOWDEFAULT);
       return (result>32);
    }
